+++ tfe/resource_tfe_registry_module_test.go
@@ -18,6 +18,14 @@ func TestAccTFERegistryModule_vcs(t *testing.T) {
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
 
+	expectedRegistryModuleAttributes := &tfe.RegistryModule{
+		Name:         getRegistryModuleName(),
+		Provider:     getRegistryModuleProvider(),
+		RegistryName: tfe.PrivateRegistry,
+		Namespace:    orgName,
+		Organization: &tfe.Organization{Name: orgName},
+	}
+
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
 			testAccPreCheck(t)
@@ -27,17 +35,30 @@ func TestAccTFERegistryModule_vcs(t *testing.T) {
 		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFERegistryModule_vcs(rInt),
+				Config:             testAccTFERegistryModule_vcs(rInt),
+				ExpectNonEmptyPlan: true,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckTFERegistryModuleExists(
-						"tfe_registry_module.foobar", orgName, registryModule),
-					testAccCheckTFERegistryModuleAttributes(registryModule, orgName),
+						"tfe_registry_module.foobar",
+						tfe.RegistryModuleID{
+							Organization: orgName,
+							Name:         expectedRegistryModuleAttributes.Name,
+							Provider:     expectedRegistryModuleAttributes.Provider,
+							RegistryName: expectedRegistryModuleAttributes.RegistryName,
+							Namespace:    orgName,
+						}, registryModule),
+					testAccCheckTFERegistryModuleAttributes(registryModule, expectedRegistryModuleAttributes),
+					testAccCheckTFERegistryModuleVCSAttributes(registryModule),
 					resource.TestCheckResourceAttr(
 						"tfe_registry_module.foobar", "organization", orgName),
 					resource.TestCheckResourceAttr(
-						"tfe_registry_module.foobar", "name", getRegistryModuleName()),
+						"tfe_registry_module.foobar", "name", expectedRegistryModuleAttributes.Name),
 					resource.TestCheckResourceAttr(
-						"tfe_registry_module.foobar", "module_provider", getRegistryModuleProvider()),
+						"tfe_registry_module.foobar", "module_provider", expectedRegistryModuleAttributes.Provider),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "namespace", expectedRegistryModuleAttributes.Namespace),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "registry_name", string(expectedRegistryModuleAttributes.RegistryName)),
 					resource.TestCheckResourceAttr(
 						"tfe_registry_module.foobar", "vcs_repo.0.display_identifier", GITHUB_REGISTRY_MODULE_IDENTIFIER),
 					resource.TestCheckResourceAttr(
@@ -69,7 +90,156 @@ func TestAccTFERegistryModule_emptyVCSRepo(t *testing.T) {
 	})
 }
 
-func TestAccTFERegistryModuleImport(t *testing.T) {
+func TestAccTFERegistryModule_nonVCSPrivateRegistryModuleWithoutRegistryName(t *testing.T) {
+	registryModule := &tfe.RegistryModule{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+
+	expectedRegistryModuleAttributes := &tfe.RegistryModule{
+		Name:         "test_module",
+		Provider:     "my_provider",
+		RegistryName: tfe.PrivateRegistry,
+		Namespace:    orgName,
+		Organization: &tfe.Organization{Name: orgName},
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config:             testAccTFERegistryModule_privateRMWithoutRegistryName(rInt),
+				ExpectNonEmptyPlan: true,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFERegistryModuleExists(
+						"tfe_registry_module.foobar",
+						tfe.RegistryModuleID{
+							Organization: orgName,
+							Name:         expectedRegistryModuleAttributes.Name,
+							Provider:     expectedRegistryModuleAttributes.Provider,
+							RegistryName: expectedRegistryModuleAttributes.RegistryName,
+							Namespace:    expectedRegistryModuleAttributes.Namespace,
+						}, registryModule),
+					testAccCheckTFERegistryModuleAttributes(registryModule, expectedRegistryModuleAttributes),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "organization", orgName),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "name", expectedRegistryModuleAttributes.Name),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "module_provider", expectedRegistryModuleAttributes.Provider),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "namespace", expectedRegistryModuleAttributes.Namespace),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "registry_name", string(expectedRegistryModuleAttributes.RegistryName)),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_nonVCSPrivateRegistryModuleWithRegistryName(t *testing.T) {
+	registryModule := &tfe.RegistryModule{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+
+	expectedRegistryModuleAttributes := &tfe.RegistryModule{
+		Name:         "another_test_module",
+		Provider:     "my_provider",
+		RegistryName: tfe.PrivateRegistry,
+		Namespace:    orgName,
+		Organization: &tfe.Organization{Name: orgName},
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config:             testAccTFERegistryModule_privateRMWithRegistryName(rInt),
+				ExpectNonEmptyPlan: true,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFERegistryModuleExists(
+						"tfe_registry_module.foobar",
+						tfe.RegistryModuleID{
+							Organization: orgName,
+							Name:         expectedRegistryModuleAttributes.Name,
+							Provider:     expectedRegistryModuleAttributes.Provider,
+							RegistryName: expectedRegistryModuleAttributes.RegistryName,
+							Namespace:    expectedRegistryModuleAttributes.Namespace,
+						}, registryModule),
+					testAccCheckTFERegistryModuleAttributes(registryModule, expectedRegistryModuleAttributes),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "organization", orgName),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "name", expectedRegistryModuleAttributes.Name),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "module_provider", expectedRegistryModuleAttributes.Provider),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "namespace", expectedRegistryModuleAttributes.Namespace),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "registry_name", string(expectedRegistryModuleAttributes.RegistryName)),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_publicRegistryModule(t *testing.T) {
+	registryModule := &tfe.RegistryModule{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	orgName := fmt.Sprintf("tst-terraform-%d", rInt)
+
+	expectedRegistryModuleAttributes := &tfe.RegistryModule{
+		Name:         "vpc",
+		Provider:     "aws",
+		RegistryName: tfe.PublicRegistry,
+		Namespace:    "terraform-aws-modules",
+		Organization: &tfe.Organization{Name: orgName},
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFERegistryModule_publicRM(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFERegistryModuleExists(
+						"tfe_registry_module.foobar",
+						tfe.RegistryModuleID{
+							Organization: orgName,
+							Name:         expectedRegistryModuleAttributes.Name,
+							Provider:     expectedRegistryModuleAttributes.Provider,
+							RegistryName: expectedRegistryModuleAttributes.RegistryName,
+							Namespace:    expectedRegistryModuleAttributes.Namespace,
+						}, registryModule),
+					testAccCheckTFERegistryModuleAttributes(registryModule, expectedRegistryModuleAttributes),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "organization", orgName),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "name", expectedRegistryModuleAttributes.Name),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "module_provider", expectedRegistryModuleAttributes.Provider),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "namespace", expectedRegistryModuleAttributes.Namespace),
+					resource.TestCheckResourceAttr(
+						"tfe_registry_module.foobar", "registry_name", string(expectedRegistryModuleAttributes.RegistryName)),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModuleImport_vcsPrivateRMDeprecatedFormat(t *testing.T) {
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
 	resource.Test(t, resource.TestCase{
@@ -81,7 +251,8 @@ func TestAccTFERegistryModuleImport(t *testing.T) {
 		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFERegistryModule_vcs(rInt),
+				Config:             testAccTFERegistryModule_vcs(rInt),
+				ExpectNonEmptyPlan: true,
 			},
 			{
 				ResourceName:        "tfe_registry_module.foobar",
@@ -93,7 +264,168 @@ func TestAccTFERegistryModuleImport(t *testing.T) {
 	})
 }
 
-func testAccCheckTFERegistryModuleExists(n, orgName string, registryModule *tfe.RegistryModule) resource.TestCheckFunc {
+func TestAccTFERegistryModuleImport_vcsPrivateRMRecommendedFormat(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+			testAccPreCheckTFERegistryModule(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config:             testAccTFERegistryModule_vcs(rInt),
+				ExpectNonEmptyPlan: true,
+			},
+			{
+				ResourceName:        "tfe_registry_module.foobar",
+				ImportState:         true,
+				ImportStateIdPrefix: fmt.Sprintf("tst-terraform-%d/%v/tst-terraform-%d/%v/%v/", rInt, "private", rInt, getRegistryModuleName(), getRegistryModuleProvider()),
+				ImportStateVerify:   true,
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModuleImport_nonVCSPrivateRM(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config:             testAccTFERegistryModule_privateRMWithRegistryName(rInt),
+				ExpectNonEmptyPlan: true,
+			},
+			{
+				ResourceName:        "tfe_registry_module.foobar",
+				ImportState:         true,
+				ImportStateIdPrefix: fmt.Sprintf("tst-terraform-%d/%v/tst-terraform-%d/%v/%v/", rInt, "private", rInt, "another_test_module", "my_provider"),
+				ImportStateVerify:   true,
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModuleImport_publicRM(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFERegistryModuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFERegistryModule_publicRM(rInt),
+			},
+			{
+				ResourceName:        "tfe_registry_module.foobar",
+				ImportState:         true,
+				ImportStateIdPrefix: fmt.Sprintf("tst-terraform-%d/%v/%v/%v/%v/", rInt, "public", "terraform-aws-modules", "vpc", "aws"),
+				ImportStateVerify:   true,
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidWithBothVCSRepoAndModuleProvider(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidWithBothVCSRepoAndModuleProvider(),
+				ExpectError: regexp.MustCompile("\"module_provider\": only one of `module_provider,vcs_repo` can be specified,\nbut `module_provider,vcs_repo` were specified."),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidRegistryName(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidRegistryName(),
+				ExpectError: regexp.MustCompile(`invalid value for registry-name. It must be either "private" or "public"`),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidWithModuleProviderAndNoName(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidWithModuleProviderAndNoName(),
+				ExpectError: regexp.MustCompile("\"module_provider\": all of `module_provider,name,organization` must be\nspecified"),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidWithModuleProviderAndNoOrganization(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidWithModuleProviderAndNoOrganization(),
+				ExpectError: regexp.MustCompile("\"module_provider\": all of `module_provider,name,organization` must be\nspecified"),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidWithNamespaceAndNoRegistryName(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidWithNamespaceAndNoRegistryName(),
+				ExpectError: regexp.MustCompile("\"namespace\": all of `namespace,registry_name` must be specified"),
+			},
+		},
+	})
+}
+
+func TestAccTFERegistryModule_invalidWithRegistryNameAndNoModuleProvider(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config:      testAccTFERegistryModule_invalidWithRegistryNameAndNoModuleProvider(),
+				ExpectError: regexp.MustCompile("\"registry_name\": all of `module_provider,registry_name` must be specified"),
+			},
+		},
+	})
+}
+func testAccCheckTFERegistryModuleExists(n string, rmID tfe.RegistryModuleID, registryModule *tfe.RegistryModule) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		tfeClient := testAccProvider.Meta().(*tfe.Client)
 
@@ -104,12 +436,6 @@ func testAccCheckTFERegistryModuleExists(n, orgName string, registryModule *tfe.
 
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("No instance ID is set")
-		}
-
-		rmID := tfe.RegistryModuleID{
-			Organization: orgName,
-			Name:         getRegistryModuleName(),
-			Provider:     getRegistryModuleProvider(),
 		}
 
 		rm, err := tfeClient.RegistryModules.Read(ctx, rmID)
@@ -127,20 +453,34 @@ func testAccCheckTFERegistryModuleExists(n, orgName string, registryModule *tfe.
 	}
 }
 
-func testAccCheckTFERegistryModuleAttributes(registryModule *tfe.RegistryModule, orgName string) resource.TestCheckFunc {
+func testAccCheckTFERegistryModuleAttributes(registryModule *tfe.RegistryModule, expectedRegistryModuleAttributes *tfe.RegistryModule) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		if registryModule.Name != getRegistryModuleName() {
+		if registryModule.Name != expectedRegistryModuleAttributes.Name {
 			return fmt.Errorf("Bad name: %s", registryModule.Name)
 		}
 
-		if registryModule.Provider != getRegistryModuleProvider() {
+		if registryModule.Provider != expectedRegistryModuleAttributes.Provider {
 			return fmt.Errorf("Bad module_provider: %s", registryModule.Provider)
 		}
 
-		if registryModule.Organization.Name != orgName {
+		if registryModule.Organization.Name != expectedRegistryModuleAttributes.Organization.Name {
 			return fmt.Errorf("Bad organization: %v", registryModule.Organization.Name)
 		}
 
+		if registryModule.RegistryName != expectedRegistryModuleAttributes.RegistryName {
+			return fmt.Errorf("Bad registry_name: %v", registryModule.RegistryName)
+		}
+
+		if registryModule.Namespace != expectedRegistryModuleAttributes.Namespace {
+			return fmt.Errorf("Bad namespace: %v", registryModule.Namespace)
+		}
+
+		return nil
+	}
+}
+
+func testAccCheckTFERegistryModuleVCSAttributes(registryModule *tfe.RegistryModule) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
 		if registryModule.VCSRepo == nil {
 			return fmt.Errorf("Bad VCS repo: %v", registryModule.VCSRepo)
 		}
@@ -189,10 +529,22 @@ func testAccCheckTFERegistryModuleDestroy(s *terraform.State) error {
 			return fmt.Errorf("No module_provider is set for registry module %s", id)
 		}
 
+		namespace := rs.Primary.Attributes["namespace"]
+		if namespace == "" {
+			return fmt.Errorf("No namespace is set for registry module %s", id)
+		}
+
+		registry_name := rs.Primary.Attributes["registry_name"]
+		if registry_name == "" {
+			return fmt.Errorf("No registry_name is set for registry module %s", id)
+		}
+
 		rmID := tfe.RegistryModuleID{
 			Organization: organization,
 			Name:         name,
 			Provider:     module_provider,
+			Namespace:    rs.Primary.Attributes["namespace"],
+			RegistryName: tfe.RegistryName(rs.Primary.Attributes["registry_name"]),
 		}
 		_, err := tfeClient.RegistryModules.Read(ctx, rmID)
 		if err == nil {
@@ -278,4 +630,112 @@ resource "tfe_oauth_client" "foobar" {
 resource "tfe_registry_module" "foobar" {
  vcs_repo {}
 }`, rInt, token)
+}
+
+func testAccTFERegistryModule_privateRMWithoutRegistryName(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+ name  = "tst-terraform-%d"
+ email = "admin@company.com"
+}
+
+resource "tfe_registry_module" "foobar" {
+	organization    = tfe_organization.foobar.id
+  module_provider = "my_provider"
+  name            = "test_module"
+ }`,
+		rInt)
+}
+
+func testAccTFERegistryModule_privateRMWithRegistryName(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+ name  = "tst-terraform-%d"
+ email = "admin@company.com"
+}
+
+resource "tfe_registry_module" "foobar" {
+	organization    = tfe_organization.foobar.id
+  module_provider = "my_provider"
+  name            = "another_test_module"
+  registry_name   = "private"
+ }`,
+		rInt)
+}
+
+func testAccTFERegistryModule_publicRM(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+ name  = "tst-terraform-%d"
+ email = "admin@company.com"
+}
+
+resource "tfe_registry_module" "foobar" {
+  organization    = tfe_organization.foobar.id
+  namespace       = "terraform-aws-modules"
+  module_provider = "aws"
+  name            = "vpc"
+  registry_name   = "public"
+ }`,
+		rInt)
+}
+
+func testAccTFERegistryModule_invalidWithBothVCSRepoAndModuleProvider() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  module_provider = "aws"
+	vcs_repo {
+		display_identifier = "hashicorp/terraform-random-module"
+		identifier         = "hashicorp/terraform-random-module"
+		oauth_token_id     = "sample-auth-token"
+	}
+ }`
+}
+
+func testAccTFERegistryModule_invalidRegistryName() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  organization    = "hashicorp"
+  module_provider = "aws"
+  name            = "eks"
+  registry_name   = "PRIVATE"
+ }`
+}
+
+func testAccTFERegistryModule_invalidWithModuleProviderAndNoName() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  organization    = "hashicorp"
+  module_provider = "aws"
+  registry_name   = "private"
+ }`
+}
+
+func testAccTFERegistryModule_invalidWithModuleProviderAndNoOrganization() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  name            = "eks"
+  module_provider = "aws"
+  registry_name   = "private"
+ }`
+}
+
+func testAccTFERegistryModule_invalidWithNamespaceAndNoRegistryName() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  organization    = "hashicorp"
+  module_provider = "aws"
+  name            = "eks"
+  namespace       = "terraform-aws-modules"
+ }`
+}
+
+func testAccTFERegistryModule_invalidWithRegistryNameAndNoModuleProvider() string {
+	return `
+resource "tfe_registry_module" "foobar" {
+  organization    = "hashicorp"
+  name            = "eks"
+  namespace       = "terraform-aws-modules"
+	registry_name   = "private"
+ }`
 }